Updated OrderCreatedEvent to send products quantities too

+++ ProductService/src/main/java/com/product/service/query/salesmetrics/SalesMetricsView.java
@@ -22,7 +22,7 @@ public class SalesMetricsView {
     @Column(name = "product_id")
     private UUID productId;
     @Column(name = "total_sold")
-    private int totalSold;
+    private long totalSold;
     @Column(name = "last_sold")
     private OffsetDateTime lastSold;
     @Column(name = "sold_last_month")
